Add more data in db & add car model and car interface

+++ F1WebApi/Interfaces/IRace.cs
@@ -4,7 +4,7 @@ public interface IRace
 {
     int Id { get; set; }
     string? WinnerName { get; set; }
-    int WinnerTime { get; set; }
+    double WinnerTime { get; set; }
     string? GrandPrix { get; set; }
     int NumberOfLaps { get; set; }
 }
